대화형 ChatGPT API 적용 및 API 요청중에 대한 표시 추가

+++ app/src/main/java/com/slee2/chatbot/data/model/Message.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "message")
 data class Message(
     @PrimaryKey(autoGenerate = true)
-    var _id: Int = 0,
+    var _id: Long = 0,
     var message: String,
     var type: Int,
 ) {
